requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
dm-tree==0.1.9
docstring_parser==0.16
einops==0.8.1
etils==1.12.1
executing==2.2.0
filelock==3.17.0
Flask==2.2.3
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.69.1
grpcio==1.70.0
h5py==3.13.0
huggingface-hub==0.29.2
idna==3.10
imageio==2.37.0
immutabledict==4.2.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipynb-py-convert==0.4.6
ipython==9.0.1
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
promise==2.3
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
simple-parsing==0.1.7
six==1.17.0
spektral==1.3.1
spicy==0.16.0
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.1
tensorflow-datasets==4.9.7
tensorflow-hub==0.16.1
tensorflow-metadata==1.16.1
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tifffile==2025.2.18
timm==1.0.15
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1+cu121
torch-geometric==2.6.1
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
validators==0.34.0
vit-keras==0.1.0
wcwidth==0.2.13
Werkzeug==2.2.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
